requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
networkx>=2.6.0
umap-learn>=0.5.0
scikit-learn>=0.24.0
imageio>=2.9.0 